add necessary platform checks
