+++ src/PerformanceTests/Utils/SymbolicLink.cs
@@ -1,13 +1,16 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Runtime.InteropServices;
 
 public class SymbolicLink
 {
     class Win32
     {
+#pragma warning disable PC003
         [DllImport("kernel32.dll")]
         [return: MarshalAs(UnmanagedType.I1)]
         public static extern bool CreateSymbolicLink(string lpSymlinkFileName, string lpTargetFileName, LinkType dwFlags);
+#pragma warning restore PC003
     }
 
     enum LinkType
@@ -18,6 +21,11 @@ public class SymbolicLink
 
     public static bool Create(string source, string destination)
     {
+        if (Environment.OSVersion.Platform != PlatformID.Win32NT)
+        {
+            return false;
+        }
+
         if (!File.Exists(source)) throw new FileNotFoundException("Source file not found.", source);
         return Win32.CreateSymbolicLink(destination, source, LinkType.File);
     }
